유진 - 수정

+++ Project_RailTraveler/src/main/java/com/railer/rt/member/MemberController.java
@@ -318,7 +318,7 @@ public class MemberController {
 	
 	@RequestMapping(value="/member/deleteMember", method= RequestMethod.GET)
 	public String deleteMember(
-			@RequestParam long userNum,
+			@RequestParam int userNum,
 			HttpSession session,
 			final RedirectAttributes reAttr) {
